Added tests for the inactivity timeout
To do so, a special TimeContext struct has been implemented that wraps the SystemTime but can be configured also to act as MockTime, so that inside unit tests it is possible to control the time (for instance, moving 60 seconds ahead with just one instruction, without really waiting for such long time).

This struct is expected to be passed by the top level of the application (main) to any component, and for this reason it needs to be included in an Arc (thread safety is already guaranteed by the usage of a Mutex to store the mock time).