Added a YAML configuration file
This file can be used to configure several parameters of the application, like the ip address to connect to, the inactivity timeout, etc.

The existing parser for command line arguments has been modified so that it can be used to get the path of the configuration file (otherwise the current directory would be used) and then to overwrite some parameters .

For instance, setting a --timeout from command line would replace the value found in the configuration file.