Implemented a basic DoS defense against P2P message flooding
Considering that malicious nodes can try to cause out of memory crashes, it is worth to have a mechanism to limit the size of queues that are involved in P2P communication, in particular the receiving and sending buffers (queue of messages).

The adopted solution is very naive, as it limits the number of messages stored regardless of their size.

This is due to the fact that MPSC channels don't have an interface to query the current size.

The best approach would be to retrieve the size of the queue and discards messages based on the real size occupied.

We need to find another channel implementation or develop something on our side.